Move FormField back to FormField.kt
